+++ backend/src/main/java/com/planmanager/www/model/eixos/EixoDTO.java
@@ -5,14 +5,14 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 
 public record EixoDTO(
-    Long id,
+    int id,
 
     @NotBlank(message = "Title is mandatory")
     @Size(min = 3, message = "Title must be at least 3 characters")
     String titulo,
     @NotNull(message = "Prefeitura ID is mandatory")
 
-    Long prefeituraId
+    int prefeituraId
 ) {
 
 }
