+++ backend/src/main/java/com/planmanager/www/model/eixos/EixoRequestDTO.java
@@ -9,7 +9,7 @@ public record EixoRequestDTO(
     @Size(min = 3, max = 100, message = "O nome do eixo deve ter entre 3 e 100 caracteres")
     String titulo,
     @NotNull(message = "A prefeitura é obrigatória")
-    Long prefeituraId
+    int prefeituraId
 ) {
 
 }
