+++ backend/src/main/java/com/planmanager/www/model/orgaos/OrgaoGestorRequestDTO.java
@@ -9,7 +9,7 @@ public record OrgaoGestorRequestDTO(
     @Size(min = 3, max = 100, message = "O nome do órgão gestor deve ter entre 3 e 100 caracteres")
     String titulo,
     @NotNull(message = "A prefeitura é obrigatória")
-    Long prefeituraId
+    int prefeituraId
 ) {
 
 }
