+++ backend/src/main/java/com/planmanager/www/model/planos/PlanoRequestDTO.java
@@ -9,7 +9,7 @@ public record PlanoRequestDTO(
     @Size(min = 3, max = 100, message = "O nome do plano deve ter entre 3 e 100 caracteres")
     String titulo,
     @NotNull(message = "A prefeitura é obrigatória")
-    Long prefeituraId
+    int prefeituraId
 ) {
 
 }
