+++ backend/src/main/java/com/planmanager/www/repositories/OrgaosGestoresRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.planmanager.www.model.orgaos.OrgaoGestor;
 
-public interface OrgaosGestoresRepository extends JpaRepository<OrgaoGestor, Long> {
+public interface OrgaosGestoresRepository extends JpaRepository<OrgaoGestor, Integer> {
 
 }
