+++ backend/src/main/java/com/planmanager/www/repositories/PropostaSnapshotRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.planmanager.www.model.propostas.PropostaSnapshot;
 
-public interface PropostaSnapshotRepository extends JpaRepository<PropostaSnapshot, Long> {
+public interface PropostaSnapshotRepository extends JpaRepository<PropostaSnapshot, Integer> {
     @Override
     public List<PropostaSnapshot> findAll();
     
-    public List<PropostaSnapshot> findByPropostaId(Long propostaId);
+    public List<PropostaSnapshot> findByPropostaId(int propostaId);
 }
